thermal: imx91: bug fix: Temperature read returns "Resource temporarily unavailable" for values below 0 degree
This issue occurs because the register DATA0[DATA_VAL] is a 16-bit signed value,
but the temperature was not being processed as such before being converted to a 32-bit value.

The fix ensures that the temperature is correctly calculated as a 16-bit signed value
before the conversion to a 32-bit format. This prevents misinterpretation of negative
temperatures and resolves the "Resource temporarily unavailable" error.

+++ drivers/thermal/imx91_thermal.c
@@ -114,7 +114,7 @@ static int imx91_tmu_get_temp(struct thermal_zone_device *tz, int *temp)
 		return -EAGAIN;
 
 	val = readl_relaxed(tmu->base + DATA0) & 0xffffU;
-	*temp = (int)val * 1000LL / 64LL;
+	*temp = (int)((int16_t)val * 1000LL / 64LL);
 	if (*temp < TMU_TEMP_LOW_LIMIT || *temp > TMU_TEMP_HIGH_LIMIT)
 		return -EAGAIN;
